Move ColorObjectDecoder & ImageObjectDecoder
Move ColorObjectDecoder & ImageObjectDecoder from drivers context
because they can operate without any specific drivers information
must not be specialized.
