+++ src/Decoders/ImageObjectDecoder.php
@@ -4,8 +4,24 @@ declare(strict_types=1);
 
 namespace Intervention\Image\Decoders;
 
-use Intervention\Image\Drivers\SpecializableDecoder;
+use Intervention\Image\Drivers\AbstractDecoder;
+use Intervention\Image\Exceptions\DecoderException;
+use Intervention\Image\Interfaces\ImageInterface;
+use Intervention\Image\Interfaces\ColorInterface;
 
-class ImageObjectDecoder extends SpecializableDecoder
+class ImageObjectDecoder extends AbstractDecoder
 {
+    /**
+     * {@inheritdoc}
+     *
+     * @see DecoderInterface::decode()
+     */
+    public function decode(mixed $input): ImageInterface|ColorInterface
+    {
+        if (!is_a($input, ImageInterface::class)) {
+            throw new DecoderException('Unable to decode input');
+        }
+
+        return $input;
+    }
 }
