+++ tests/Unit/Drivers/Imagick/Decoders/ImageObjectDecoderTest.php
@@ -6,12 +6,12 @@ namespace Intervention\Image\Tests\Unit\Drivers\Imagick\Decoders;
 
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\RequiresPhpExtension;
-use Intervention\Image\Drivers\Imagick\Decoders\ImageObjectDecoder;
+use Intervention\Image\Decoders\ImageObjectDecoder;
 use Intervention\Image\Image;
 use Intervention\Image\Tests\ImagickTestCase;
 
 #[RequiresPhpExtension('imagick')]
-#[CoversClass(\Intervention\Image\Drivers\Imagick\Decoders\ImageObjectDecoder::class)]
+#[CoversClass(\Intervention\Image\Decoders\ImageObjectDecoder::class)]
 final class ImageObjectDecoderTest extends ImagickTestCase
 {
     public function testDecode(): void
